WLED_RGB_ALEXA_EFFECTS
WLED_RGB_ALEXA_EFFECTS: this uses the "node-red-contrib-alexa-local" Alexa-Local node to communicate to Alexa. NO extra Alexa Skills required and NO account linking required to use this node. How this work is you place the Alexa-Local in your flow and give it a name, then tell alexa to search for new devices. once she finds them they are treated as lights. then you just say "Alexa Preset 1 on" or "Alexa turn on Preset 1". The first 3 nodes in this flow also have extra nodes called "Solid", "Halloween", and "Christmas". I have set the presets in WLED to be the 3 previously named presets and so you can then say "Alexa Christmas Lights on" or "Alexa Turn on Solid Lights" or "Alexa Solid Lights on". you can change them all to named devices or leave them at "Preset 1" and so on.